Added error handling for address geolocation
Now informs the user if there are connection problems when getting a latlon from address, and successfully aborts. Switched some static info from Main to CurrentStates, which stores static info now. Reworked map marker drawing and marker addition and removal, and connected the functionality of the View Nearby X buttons in the map gui so they now add the nearby points to the map. Ensured the map did not have duplicate points by enabling hashsets of Route, RetailLocation, and WifiLocation objects that are now used to store the current expected markers.