Mainly adding progress bars to imports and async geocoding for CSV retailers
Added a ProgressPopup that occurs whenever a csv is imported, has a progress bar and a text field that can be changed, used by the CSV importers to indicate the current status of the import (success, fail, total, todo). CSV importing was greatly refractored into a single class that relies on new interfaces implemented by the three datahandlers to reduce code repitition. This CSV importer uses callbacks to identify results rather than a function returning a boolean as it enables async importing, which is especially important in the case of Retailers, which require geocoding. Useful for any further implementations that have long single run times that can be asynced. GeocodeOutcome was introduced, which uses a callback to return a result from the geocode api. This class also takes some arguments, specifically a record object which is the standard string list parsed from the CSV, and two callbacks, one to call back to with the result (geoCallback), and one to pass back or return false on if failure. This is used specifically for the CSV importing and could be abstracted or overloaded if the second callback was not needed, which it wouldn't be if we were using this not linked to the CSVImporter class. As is, these results are passed to the callback so it can then use the results to add to the database, which then uses the callback it was passed to return a success value. Many tests and jdocs were broken, they will be updated tomorrow but I wanted to push this func even with them broken. There are still errors when incorrect CSVs are imported and such which were previously handled, so will fix that tomorrow too